tambah atribut stok

+++ app/Http/Controllers/TransaksiController.php
@@ -6,48 +6,70 @@ use App\Models\Keranjang;
 use App\Models\Produk;
 use App\Models\Transaksi;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 
 class TransaksiController extends Controller
 {
-    public function store($id) {
-        $id = Auth::user()->id;
-        $keranjang = Keranjang::all()->where('pembeli_id', $id);
-        
-        foreach($keranjang as $keranj) {
-            $count = Keranjang::where('produk_id', '=', $keranj->produk->id)->count();
-            $totalHarga = $keranj->harga * $count;
+    public function store(Request $request) {
+        $provinsi = $request->get('Province');
+        $kota = $request->get('city');
+        $alamat = $request->get('addressOne');
+        $nohp = $request->get('mobile');
+        $jumlahBarang = $request->get('counts');
 
-            $product = new Transaksi([
-                "total_harga" => $totalHarga,
-                "jumlah_barang" => $count,
-                "alamat" => 'Balikpapan',
-                "pembeli_id" => $id,
-                "penjual_id" => $keranj->produk->penjual_id,
-                "produk_id" => $keranj->produk->id,
-            ]); 
-            $product->save();
-            
-            if(Transaksi::where('pembeli_id', '=', $id)->where('produk_id', '=', $keranj->produk->id)->where('jumlah_barang', '=', $count)->count() > 1) {
-                $product->delete();
+        if($jumlahBarang > 0) {
+            $id = Auth::user()->id;
+            $carts = Keranjang::all()->where('pembeli_id', $id);
+            $alamatLengkap = $provinsi.', '.$kota.'. '.$alamat.'; '.$nohp;
+
+            foreach($carts as $cart) {
+                $produk = Produk::findOrFail($cart->produk->id);
+
+                $sisaBarang = $produk->stok - $cart->jumlah_barang;
+                // if($sisaBarang == 0) {
+                //     $produk->delete();
+                // } else {
+                // }
+                $produk->update(['stok' => $sisaBarang]);
+
+                $transaksi = new Transaksi([
+                    "total_harga" => $cart->total_harga,
+                    "jumlah_barang" => $cart->jumlah_barang,
+                    "alamat" => $alamatLengkap,
+                    "pembeli_id" => $id,
+                    "penjual_id" => $cart->produk->penjual_id,
+                    "produk_id" => $cart->produk->id,
+                ]); 
+                $transaksi->save();
             }
+            Keranjang::where('pembeli_id', $id)->delete();
+            return redirect('/success');
+        } else {
+            return redirect('/keranjang');
         }
-        Keranjang::where('pembeli_id', $id)->delete();
-
-        return redirect('/success');
     }
     
-    public function storeKeranjang(Produk $products) {
+    public function storeProduk(Produk $product) {
+        $sisaBarang = $product->stok - 1;
+        $product->update(['stok' => $sisaBarang]);
+        
         $transaksi = new Transaksi([
-            "total_harga" => $products->harga,
+            "total_harga" => $product->harga,
             "jumlah_barang" => '1',
-            "alamat" => 'Samarinda',
+            "alamat" => Auth::user()->address,
             "pembeli_id" => Auth::user()->id,
-            "penjual_id" => $products->penjual_id,
-            "produk_id" => $products->id,
+            "penjual_id" => $product->penjual_id,
+            "produk_id" => $product->id,
         ]); 
         $transaksi->save();
         
-        return redirect("/show/$products->id")->with('success', 'Produk berhasil di Checkout!');
+        return redirect("/show/$product->id")->with('success', 'Produk berhasil di Checkout!');
+    }
+
+    public function daftarTransaksi() {
+        $transaksi = Transaksi::all()->where('pembeli_id', Auth::user()->id);
+
+        return view('pembeli.checkout', ["transactions" => $transaksi]);
     }
 
     public function penjual() {
